refactor(episode edit page): break 602-line page into hook + parts
- useEpisodeEdit hook owns slug resolution, data loading (phrases + extractions
  with live counts), and extraction/episode delete + phrase-edit navigation
- EpisodeEditHeader, EpisodeSettingsCard, ExtractionsSection, ExtractionRow
  colocated under edit/components/
- New shared StatCard primitive replaces the 4 repeated metric cards
- page.tsx stays the route entry, now a thin orchestrator
- Typed extractions as EpisodeExtraction[] (removes one any; lint 50->49)

Verified: tsc clean, lint 0 errors, build succeeds. No behavior change.
